Updating profile setup test journey to force user through expected path
Normally after a user has set up their profile already then redirecting
to the homepage will not force them into the next step of the profile
setup journey - the profile checker will determine that they already
have an organisation, interests, grade, etc set.

In order to force a given test user through this journey, in lieu of
resetting all the user data in between individual thread completions we
can instead set the 'originalUrl' form parameter to point to the next
step in the profile setup flow, which will redirect the user to the
intended form.